add eslint and import sorting (not working yet)
