+++ .eslintrc
@@ -1,0 +1,33 @@
+{
+    "parser": "@typescript-eslint/parser",
+    "extends": [
+        "eslint:recommended",
+        "prettier"
+    ],
+    "env": {
+        "es6": true
+    },
+    "plugins": [
+        "svelte3",
+        "@typescript-eslint",
+        "simple-import-sort"
+    ],
+    "overrides": [
+        {
+            "files": [
+                "*.svelte"
+            ],
+            "processor": "svelte3/svelte3"
+        }
+    ],
+    "rules": {
+        "simple-import-sort/imports": "error",
+        "simple-import-sort/exports": "error",
+        "import/first": "error",
+        "import/newline-after-import": "error",
+        "import/no-duplicates": "error"
+    },
+    "settings": {
+        "svelte3/typescript": true,
+    }
+}
