+++ .prettierrc
@@ -8,5 +8,5 @@
     "arrowParens": "always",
     "endOfLine": "lf",
     "embeddedLanguageFormatting": "auto",
-    "singleAttributePerLine": true
+    "singleAttributePerLine": true,
 }
