+++ package.json
@@ -1,31 +1,35 @@
 {
-  "name": "rmx-web",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.0.2",
-    "@trivago/prettier-plugin-sort-imports": "^4.0.0",
-    "@tsconfig/svelte": "^3.0.0",
-    "sass": "^1.54.9",
-    "svelte": "^3.49.0",
-    "svelte-check": "^3.0.2",
-    "svelte-preprocess": "^5.0.0",
-    "tslib": "^2.4.0",
-    "typescript": "^4.6.4",
-    "vite": "^4.0.4"
-  },
-  "dependencies": {
-    "axios": "^1.2.2",
-    "feather-icons": "^4.29.0",
-    "fuzzysort": "^2.0.4",
-    "notiflix": "^3.2.6",
-    "svelte-navigator": "^3.2.2"
-  }
+    "name": "rmx-web",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "scripts": {
+        "dev": "vite",
+        "build": "vite build",
+        "preview": "vite preview",
+        "check": "svelte-check --tsconfig ./tsconfig.json"
+    },
+    "devDependencies": {
+        "@sveltejs/vite-plugin-svelte": "^2.0.2",
+        "@tsconfig/svelte": "^3.0.0",
+        "@typescript-eslint/eslint-plugin": "^5.50.0",
+        "@typescript-eslint/parser": "^5.50.0",
+        "eslint-config-prettier": "^8.6.0",
+        "eslint-plugin-simple-import-sort": "^10.0.0",
+        "eslint-plugin-svelte3": "^4.0.0",
+        "sass": "^1.54.9",
+        "svelte": "^3.49.0",
+        "svelte-check": "^3.0.2",
+        "svelte-preprocess": "^5.0.0",
+        "tslib": "^2.4.0",
+        "typescript": "^4.6.4",
+        "vite": "^4.0.4"
+    },
+    "dependencies": {
+        "axios": "^1.2.2",
+        "feather-icons": "^4.29.0",
+        "fuzzysort": "^2.0.4",
+        "notiflix": "^3.2.6",
+        "svelte-navigator": "^3.2.2"
+    }
 }
