+++ src/lib/components/global/Icon.svelte
