+++ src/lib/types/websocket.ts
