+++ src/pages/home/modals/CreateSessionModal.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-    import type { CreateJamData, GetJamData } from '../../../models/jam';
-    import Modal from '../../../lib/components/Modal.svelte';
     import type { AxiosError } from 'axios';
-    import { api } from '../../../api/api';
-    import { Failure, Success } from '../../../lib/notify/notify';
     import { navigate } from 'svelte-navigator';
+    import type { CreateJamData, GetJamData } from 'src/lib/types/jam';
+    import Modal from 'src/lib/components/global/Modal.svelte';
+    import { Failure, Success } from 'src/lib/notify/notify';
+    import { api } from 'src/api/api';
 
     export let closeFunc: Function;
