+++ src/store/user.ts
@@ -1,4 +1,4 @@
-import type { User } from 'src/models/user';
+import type { User } from 'src/lib/types/user';
 import { writable } from 'svelte/store';
 
 export const UserStore = writable<User>({
